Fix - Support Overlay making it impossible to clic on pages on small windows
Cause : the overlay was taking all the windows dimenson but mainly it was the popup fault.

I forgot the IsVisible property that was in the .cs code of the popup but not on the XAML code, so the opacity of 0 was making it invisible but always clikable. So we didn't see it but it dit intercept click : this explain why on mobile when we were on the conversationsPages, a click at the top on the Support Conversation was triggering a faq question that was only in the chat popup.

Now no only the overlay is not full screen, it adapt to the size taken in the ui either of the button or the popup